sanitize, fix routes
