+++ apps/website/app/api/llm/anthropic/chat/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest } from "next/server";
-import cors from "../../../../lib/cors";
+import cors from "../../../../../lib/cors";
 
 type Message = {
   role: string;
@@ -12,6 +12,11 @@ type Settings = {
   temperature: number;
 };
 
+type AnthropicUsage = {
+  input_tokens: number;
+  output_tokens: number;
+};
+
 type RequestBody = {
   documents: Message[];
   passphrase?: string;
@@ -20,6 +25,7 @@ type RequestBody = {
 
 const CONTENT_TYPE_JSON = "application/json";
 const CONTENT_TYPE_TEXT = "text/plain";
+const ANTHROPIC_API_VERSION = "2023-06-01";
 
 export const runtime = "nodejs";
 export const preferredRegion = "auto";
@@ -31,17 +37,16 @@ export async function POST(request: NextRequest): Promise<Response> {
     const { documents: messages, settings } = requestData;
     const { model, maxTokens, temperature } = settings;
 
-    // Get API key from environment variable
-    const apiKey = process.env.OPENAI_API_KEY;
+    const apiKey = process.env.ANTHROPIC_API_KEY;
 
     if (!apiKey) {
-      console.error("OPENAI_API_KEY environment variable is not set");
+      console.error("ANTHROPIC_API_KEY environment variable is not set");
       return cors(
         request,
         new Response(
           JSON.stringify({
             error:
-              "API key not configured. Please set the OPENAI_API_KEY environment variable in your Vercel project settings.",
+              "API key not configured. Please set the ANTHROPIC_API_KEY environment variable in your Vercel project settings.",
           }),
           {
             status: 500,
@@ -51,24 +56,21 @@ export async function POST(request: NextRequest): Promise<Response> {
       );
     }
 
-    const url = "https://api.openai.com/v1/chat/completions";
+    const url = "https://api.anthropic.com/v1/messages";
 
     const body = JSON.stringify({
       model: model,
+      max_tokens: maxTokens,
       messages: messages,
-      ...(temperature !== undefined && {
-        temperature: temperature,
-        max_tokens: maxTokens,
-      }),
+      temperature: temperature,
     });
-
-    console.log(`Making request to OpenAI API with model: ${model}`);
 
     const response = await fetch(url, {
       method: "POST",
       headers: {
         "Content-Type": CONTENT_TYPE_JSON,
-        Authorization: `Bearer ${apiKey}`,
+        "x-api-key": apiKey,
+        "anthropic-version": ANTHROPIC_API_VERSION,
       },
       body,
     });
@@ -76,12 +78,12 @@ export async function POST(request: NextRequest): Promise<Response> {
     const responseData = await response.json();
 
     if (!response.ok) {
-      console.error("OpenAI API error:", responseData);
+      console.error("Anthropic API error:", responseData);
       return cors(
         request,
         new Response(
           JSON.stringify({
-            error: `OpenAI API error: ${responseData.error?.message || "Unknown error"}`,
+            error: `Anthropic API error: ${responseData.error?.message || "Unknown error"}`,
           }),
           {
             status: response.status,
@@ -91,16 +93,19 @@ export async function POST(request: NextRequest): Promise<Response> {
       );
     }
 
-    const replyText = responseData.choices?.[0]?.message?.content;
+    const replyText = responseData.content?.[0]?.text;
 
     if (!replyText) {
-      console.error("Invalid response format from OpenAI API:", responseData);
+      console.error(
+        "Invalid response format from Anthropic API:",
+        responseData,
+      );
       return cors(
         request,
         new Response(
           JSON.stringify({
             error:
-              "Invalid response format from OpenAI API. Check server logs for details.",
+              "Invalid response format from Anthropic API. Check server logs for details.",
           }),
           {
             status: 500,
